fully (hope so) implement bresenhams algorithm (with floating points)

+++ primitives.h
@@ -12,7 +12,7 @@
 void prim_fill(uint *canvas, size_t width, size_t height, uint color);
 void prim_draw_rect(uint *canvas, size_t width, size_t height, size_t x1, size_t y1, size_t x2, size_t y2, 
 	uint color);
-void prim_draw_line(uint *canvas, size_t width, size_t height, size_t x1, size_t y1, size_t x2, size_t y2,
+void prim_draw_line(uint *canvas, size_t width, size_t height, int x1, int y1, int x2, int y2,
 	uint color);
 void prim_save_to_ppm_file(uint *canvas, size_t width, size_t height, const char *file_path);
